yahoo!
for who passed all the courses in the highest level in the program,
course table doesn't display nothing since the student do not need to
take any more.
for who failed some courses, course table have course in current level
and next level.
for who in the highest level and failed some will get only the courses
in the level on the course table